+++ R/read_grid.R
@@ -2,7 +2,7 @@
 #'
 #' @description
 #' Results of the AOP project are spatially aggregated on a hexagonal grid based
-#' on the global H3 index at resolution 8, with a size of 357 meters (short
+#' on the global H3 index at resolution 9, with a size of 357 meters (short
 #' diagonal) and an area of 0.74 km2. More information about H3 at
 #' \url{https://h3geo.org/docs/core-library/restable/}.  See documentation
 #' 'Details' for the data dictionary.
@@ -23,13 +23,13 @@
 #'
 #' @export
 #' @family spatial data functions
-#' @examples \donttest{
+#' @examples \dontrun{ if (interactive()) {
 #' # Read spatial grid of a single city
 #' nat <- read_grid(city = 'Natal', showProgress = FALSE)
 #'
 #' # Read spatial grid of all cities in the project
 #' # all <- read_grid(city = 'all', showProgress = FALSE)
-#'}
+#'}}
 read_grid <- function(city, showProgress = FALSE){
 
   # checks
@@ -38,10 +38,17 @@ read_grid <- function(city, showProgress = FALSE){
   # Get metadata with data url addresses
   temp_meta <- select_metadata(t="grid", c=city)
 
+  # check if download failed
+  if (is.null(temp_meta)) { return(invisible(NULL)) }
+
   # list paths of files to download
   file_url <- as.character(temp_meta$download_path)
 
   # download files
   aop_sf <- download_data(file_url, progress_bar = showProgress)
+
+  # check if download failed
+  if (is.null(aop_sf)) { return(invisible(NULL)) }
+
   return(aop_sf)
 }