+++ R/read_landuse.R
@@ -4,7 +4,7 @@
 #' Download data on the spatial distribution of population, schools and
 #' healthcare facilities at a fine spatial resolution for cities included in the
 #' study. The data comes aggregated on a hexagonal grid based on the global H3
-#' index at resolution 8, with a size of 357 meters (short diagonal) and an area
+#' index at resolution 9, with a size of 357 meters (short diagonal) and an area
 #' of 0.74 km2. More information about H3 at \url{https://h3geo.org/docs/core-library/restable/}.
 #'
 #' See documentation 'Details' for the data dictionary.
@@ -52,14 +52,14 @@
 #'
 #' @export
 #' @family land use data functions
-#' @examples \donttest{
+#' @examples \dontrun{ if (interactive()) {
 #' # a single city
 #' bho <- read_landuse(city = 'Belo Horizonte', year = 2019, showProgress = FALSE)
 #' bho <- read_landuse(city = 'bho', year = 2019, showProgress = FALSE)
 #'
 #' # all cities
 #' all <- read_landuse(city = 'all', year = 2019)
-#'}
+#'}}
 read_landuse <- function(city='bel', year = 2019, geometry = FALSE, showProgress = TRUE){
 
   # checks
@@ -71,11 +71,19 @@ read_landuse <- function(city='bel', year = 2019, geometry = FALSE, showProgress
                                c=city,
                                y=year)
 
+  # check if download failed
+  if (is.null(temp_meta)) { return(invisible(NULL)) }
+
+  message(paste0("Downloading land use data from year ", year))
+
   # list paths of files to download
   file_url <- as.character(temp_meta$download_path)
 
   # download files
   aop_landuse <- download_data(file_url, progress_bar = showProgress)
+
+  # check if download failed
+  if (is.null(aop_landuse)) { return(invisible(NULL)) }
 
   # Download and merge population data
   aop_population <- read_population(city=city, showProgress = showProgress)