+++ R/utils.R
@@ -75,7 +75,7 @@ select_year_input <- function(temp_meta=temp_meta, year=NULL){
                                    paste(unique(temp_meta$year),collapse = " "))) }
 
   # invalid input
-  else if (year %in% temp_meta$year){ message(paste0("Using year ", year))
+  else if (year %in% temp_meta$year){
                                   temp_meta <- temp_meta[ temp_meta$year %in% year, ]
                                   return(temp_meta) }
 
@@ -140,7 +140,11 @@ select_mode_input <- function(temp_meta=temp_meta, mode=NULL){
 select_metadata <- function(t=NULL, c=NULL, y=NULL, m=NULL){
 
 # download metadata
-  metadata <- as.data.frame(download_metadata())
+  metadata <- as.data.frame(aopdata::download_metadata())
+
+  # check if download failed
+  msg <- "Problem connecting to data server. Please try it again in a few minutes."
+  if (nrow(metadata)==0) { message(msg); return(invisible(NULL)) }
 
   # Select data type
   temp_meta <- subset(metadata, type == t)
@@ -149,12 +153,12 @@ select_metadata <- function(t=NULL, c=NULL, y=NULL, m=NULL){
   temp_meta <- select_city_input(temp_meta, city=c)
 
   # select year input
-  if(t %in% c('access','landuse', 'land_use', 'population')){
+  if (t %in% c('access','landuse', 'land_use', 'population')) {
     temp_meta <- select_year_input(temp_meta, year=y)
   }
 
   # select mode input
-  if(t=='access'){
+  if (t=='access') {
     temp_meta <- select_mode_input(temp_meta, mode=m)
     }
 
@@ -189,14 +193,16 @@ download_data <- function(file_url, progress_bar = showProgress){
     temps <- paste0(tempdir(),"/", unlist(lapply(strsplit(file_url,"/"),tail,n=1L)))
 
     # check if file has not been downloaded already. If not, download it
-    if (!file.exists(temps)) {
+    if (!file.exists(temps) | file.info(temps)$size == 0) {
 
       # test server connection
-      check_connection(file_url[1])
+      check_con <- check_connection(file_url[1])
+      if(is.null(check_con) | isFALSE(check_con)){ return(invisible(NULL)) }
 
       # download data
       httr::GET(url=file_url, httr::progress(), httr::write_disk(temps, overwrite = T))
     }
+
 
     # load gpkg to memory
     temp_sf <- load_data(file_url, temps)
@@ -209,10 +215,11 @@ download_data <- function(file_url, progress_bar = showProgress){
     temps <- paste0(tempdir(),"/", unlist(lapply(strsplit(file_url,"/"),tail,n=1L)))
 
     # check if file has not been downloaded already. If not, download it
-    if (!file.exists(temps)) {
+    if (!file.exists(temps) | file.info(temps)$size == 0) {
 
       # test server connection
-      check_connection(file_url[1])
+      check_con <- check_connection(file_url[1])
+      if(is.null(check_con) | isFALSE(check_con)){ return(invisible(NULL)) }
 
       # download data
       httr::GET(url=file_url, httr::write_disk(temps, overwrite = T))
@@ -234,7 +241,8 @@ download_data <- function(file_url, progress_bar = showProgress){
     pb <- utils::txtProgressBar(min = 0, max = total, style = 3)
 
     # test server connection
-    check_connection(file_url[1])
+    check_con <- check_connection(file_url[1])
+    if(is.null(check_con) | isFALSE(check_con)){ return(invisible(NULL)) }
 
     # download files
     lapply(X=file_url, function(x){
@@ -243,7 +251,7 @@ download_data <- function(file_url, progress_bar = showProgress){
       temps <- paste0(tempdir(),"/", unlist(lapply(strsplit(x,"/"),tail,n=1L)))
 
       # check if file has not been downloaded already. If not, download it
-      if (!file.exists(temps)) {
+      if (!file.exists(temps) | file.info(temps)$size == 0) {
         i <- match(c(x),file_url)
         httr::GET(url=x, #httr::progress(),
                   httr::write_disk(temps, overwrite = T))
@@ -264,7 +272,8 @@ download_data <- function(file_url, progress_bar = showProgress){
   else if(length(file_url) > 1 & progress_bar == FALSE) {
 
     # test server connection
-    check_connection(file_url[1])
+    check_con <- check_connection(file_url[1])
+    if(is.null(check_con) | isFALSE(check_con)){ return(invisible(NULL)) }
 
     # download files
     lapply(X=file_url, function(x){
@@ -273,7 +282,7 @@ download_data <- function(file_url, progress_bar = showProgress){
       temps <- paste0(tempdir(),"/", unlist(lapply(strsplit(x,"/"),tail,n=1L)))
 
       # check if file has not been downloaded already. If not, download it
-      if (!file.exists(temps)) {
+      if (!file.exists(temps) | file.info(temps)$size == 0) {
         i <- match(c(x),file_url)
         httr::GET(url=x, #httr::progress(),
                   httr::write_disk(temps, overwrite = T))
@@ -406,29 +415,54 @@ aop_merge <- function(aop_landuse, aop_access){
 #' Check internet connection with Ipea server
 #'
 #' @description
-#' Checks if there is internet connection to Ipea server to download aop data.
+#' Checks if there is internet connection with Ipea server to download aop data.
 #'
 #' @param file_url A string with the file_url address of an aop dataset
 #'
-#' @return Logic `TRUE or `FALSE`.
+#' @return Logical. `TRUE` if url is working, `FALSE` if not.
 #'
 #' @export
 #' @family support functions
 #'
 check_connection <- function(file_url = 'https://www.ipea.gov.br/geobr/aopdata/metadata/metadata.csv'){
 
-  # check internet connection
-  if (!curl::has_internet()) {
-    message("No internet connection.")
-    return(invisible(NULL))
+  # file_url <- 'https://google.com/'               # ok
+  # file_url <- 'https://www.google.com:81/'   # timeout
+  # file_url <- 'https://httpbin.org/status/300' # error
+
+  # check if user has internet connection
+  if (!curl::has_internet()) { message("No internet connection.")
+    return(FALSE)
   }
 
-  # test server connection
-  if (! crul::ok(file_url, verbose=FALSE) ) {
-    message("Problem connecting to data server. Please try aopdata again in a few minutes.")
-    return(invisible(NULL))
-    }
-}
+  # message
+  msg <- "Problem connecting to data server. Please try it again in a few minutes."
 
+  # test server connection
+  x <- try(silent = TRUE,
+           httr::GET(file_url, # timeout(5),
+                     config = httr::config(ssl_verifypeer = FALSE)))
+  # link offline
+  if (class(x)=="try-error") {
+    message( msg )
+    return(FALSE)
+  }
+
+  # link working fine
+  else if ( identical(httr::status_code(x), 200L)) {
+    return(TRUE)
+  }
+
+  # link not working or timeout
+  else if (! identical(httr::status_code(x), 200L)) {
+    message(msg )
+    return(FALSE)
+
+  } else if (httr::http_error(x) == TRUE) {
+    message(msg)
+    return(FALSE)
+  }
+
+}
 
 # nocov end